requirements.txt
google_api_python_client==2.121.0
httplib2==0.22.0
httpx==0.27.0
loguru==0.7.2
oauth2client==4.1.3
pydantic==2.6.4
pydantic_settings==2.2.1
python_dateutil==2.8.2
SQLAlchemy==2.0.28
sqlmodel==0.0.16
